Chg: Moved thread Manager to scarlett_os.utility.thread.ManagerThread
